// src/vendor/deno.land/std@0.165.0/node/_stream.d.ts
export class Stream extends EventEmitter {
  pipe<T extends WritableStream>(
    destination: T,
    options?: {
      end?: boolean | undefined;
    },
  ): T;
  constructor(opts?: ReadableOptions);
}

interface StreamOptions<T extends Stream> extends Abortable {
  emitClose?: boolean | undefined;
  highWaterMark?: number | undefined;
  objectMode?: boolean | undefined;
  construct?(this: T, callback: (error?: Error | null) => void): void;
  destroy?(
    this: T,
    error: Error | null,
    callback: (error: Error | null) => void,
  ): void;
  autoDestroy?: boolean | undefined;
}
export interface ReadableOptions extends StreamOptions<Readable> {
  encoding?: BufferEncoding | undefined;
  read?(this: Readable, size: number): void;
}
// ...
export interface WritableOptions extends StreamOptions<Writable> {
  decodeStrings?: boolean | undefined;
  defaultEncoding?: BufferEncoding | undefined;
  write?(
    this: Writable,
    chunk: any,
    encoding: BufferEncoding,
    callback: (error?: Error | null) => void,
  ): void;
  writev?(
    this: Writable,
    chunks: Array<{
      chunk: any;
      encoding: BufferEncoding;
    }>,
    callback: (error?: Error | null) => void,
  ): void;
  final?(this: Writable, callback: (error?: Error | null) => void): void;
}